requirements.txt
agent-framework
azure-identity
